backend/requirements.txt
Flask==2.2.5
flask-cors==3.0.10
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
nltk==3.8.1
scikit-learn==1.2.2
pandas==2.1.1
mysql-connector-python==8.1.0
Werkzeug==2.2.3
